pyproject.toml
[project]
name = "parser-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Ronevych Dmytro",email = "dmytro874@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "django-filter (>=25.2,<26.0)",
    "markdown (>=3.10,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
